Merge develop to master

+++ test/ChessboardTest.cpp
@@ -13,14 +13,14 @@ TEST(ChessBoardTest, BasicTest) {
       ASSERT_TRUE(squareEmpty);
     }
   }
-  ASSERT_EQ(board.figures()->size(), 0);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 0);
 }
 
 TEST(ChessBoardTest, AddFigureTest) {
   auto board = std::make_shared<Chessboard>();
-  ASSERT_EQ(board->figures()->size(), 0);
+  ASSERT_EQ(static_cast<int>(board->figures()->size()), 0);
   ASSERT_TRUE(board->addFigure(std::make_shared<RookFigure>(&board->squares()->at(0)[0])));
-  ASSERT_EQ(board->figures()->size(), 1);
+  ASSERT_EQ(static_cast<int>(board->figures()->size()), 1);
   ASSERT_FALSE(board->squares()->at(0)[0].isEmpty());
   ASSERT_EQ(board->figures()->begin()->get()->square()->coordinates()->column(), 0);
   ASSERT_EQ(board->figures()->begin()->get()->square()->coordinates()->row(), 0);
@@ -28,34 +28,34 @@ TEST(ChessBoardTest, AddFigureTest) {
 
 TEST(ChessBoardTest, AddSameFigureTest) {
   Chessboard board;
-  ASSERT_EQ(board.figures()->size(), 0);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 0);
   auto rook = std::make_shared<RookFigure>(&board.squares()->at(0)[0]);
   ASSERT_TRUE(board.addFigure(rook));
-  ASSERT_EQ(board.figures()->size(), 1);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 1);
   ASSERT_FALSE(board.squares()->at(0)[0].isEmpty());
   ASSERT_EQ(board.figures()->begin()->get()->square()->coordinates()->column(), 0);
   ASSERT_EQ(board.figures()->begin()->get()->square()->coordinates()->row(), 0);
   ASSERT_FALSE(board.addFigure(rook));
-  ASSERT_EQ(board.figures()->size(), 1);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 1);
 }
 
 TEST(ChessBoardTest, AddFigureToOccupiedSquareTest) {
   Chessboard board;
-  ASSERT_EQ(board.figures()->size(), 0);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 0);
   ASSERT_TRUE(board.addFigure(std::make_shared<RookFigure>(&board.squares()->at(0)[0])));
-  ASSERT_EQ(board.figures()->size(), 1);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 1);
   ASSERT_FALSE(board.squares()->at(0)[0].isEmpty());
   ASSERT_THROW(board.addFigure(std::make_shared<RookFigure>(&board.squares()->at(0)[0])), std::logic_error);
-  ASSERT_EQ(board.figures()->size(), 1);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 1);
 }
 
 TEST(ChessBoardTest, AddMultipleFiguresTest) {
   Chessboard board;
-  ASSERT_EQ(board.figures()->size(), 0);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 0);
   auto figuresIterator = board.figures();
   for (uint8_t i = 0; i < 8; ++i) {
     ASSERT_TRUE(board.addFigure(std::make_shared<RookFigure>(&board.squares()->at(i)[i])));
-    ASSERT_EQ(board.figures()->size(), i + 1);
+    ASSERT_EQ(static_cast<int>(board.figures()->size()), i + 1);
     ASSERT_EQ(board.figures()->begin()->get()->square()->coordinates()->column(), i);
     ASSERT_EQ(board.figures()->begin()->get()->square()->coordinates()->row(), i);
     ++figuresIterator;
@@ -64,7 +64,7 @@ TEST(ChessBoardTest, AddMultipleFiguresTest) {
 
 TEST(ChessBoardTest, MoveFigureTest) {
   Chessboard board;
-  ASSERT_EQ(board.figures()->size(), 0);
+  ASSERT_EQ(static_cast<int>(board.figures()->size()), 0);
   ASSERT_TRUE(board.addFigure(std::make_shared<RookFigure>(&board.squares()->at(0)[0])));
   ASSERT_FALSE(board.squares()->at(0)[0].isEmpty());
   ASSERT_NO_THROW(board.figures()->begin()->get()->move());
